Add BasicCurrencyAdapter.
